Fix mobile rendering

+++ components/LandingEffect/CheckerEffect.frag
@@ -1,4 +1,4 @@
-precision mediump float;
+precision highp float;
 varying vec3 vVertexPosition;
 varying vec2 vTextureCoord;
 
@@ -91,7 +91,6 @@ float noiseChecker(ivec2 currentCell, vec2 noiseScale, vec2 noiseOffset) {
 }
 
 void main() {
-
   vec2 textureCoord = vTextureCoord;
   vec3 vertexPosition = vVertexPosition;
   float aspectRatio = uResolution.y / uResolution.x;
@@ -130,6 +129,7 @@ void main() {
   // ==============================================
   // Introduce mouse influence to the checker map
   // ==============================================
+
 
   // mouse size in pixel
   float mouseSize = uCheckerSize * uCursorSize;
